+++ packages/backend/server/src/plugins/calendar/providers/def.ts
@@ -2,12 +2,7 @@ import { Inject, Injectable, Logger } from '@nestjs/common';
 import type { CalendarAccount } from '@prisma/client';
 
 import { CalendarProviderRequestError, Config, OnEvent } from '../../../base';
-import type { CalendarProviderFactory } from './factory';
-
-export enum CalendarProviderName {
-  Google = 'google',
-  CalDAV = 'caldav',
-}
+import { CalendarProviderFactory, CalendarProviderName } from './factory';
 
 export interface CalendarProviderTokens {
   accessToken: string;
