+++ packages/backend/server/src/plugins/calendar/providers/factory.ts
@@ -1,12 +1,20 @@
 import { Injectable, Logger } from '@nestjs/common';
 
-import type { CalendarProvider } from './def';
-import { CalendarProviderName } from './def';
+export enum CalendarProviderName {
+  Google = 'google',
+  CalDAV = 'caldav',
+}
+
+export interface CalendarProviderRef {
+  provider: CalendarProviderName;
+}
 
 @Injectable()
-export class CalendarProviderFactory {
+export class CalendarProviderFactory<
+  TProvider extends CalendarProviderRef = CalendarProviderRef,
+> {
   private readonly logger = new Logger(CalendarProviderFactory.name);
-  readonly #providers = new Map<CalendarProviderName, CalendarProvider>();
+  readonly #providers = new Map<CalendarProviderName, TProvider>();
 
   get providers() {
     return Array.from(this.#providers.keys());
@@ -16,12 +24,12 @@ export class CalendarProviderFactory {
     return this.#providers.get(name);
   }
 
-  register(provider: CalendarProvider) {
+  register(provider: TProvider) {
     this.#providers.set(provider.provider, provider);
     this.logger.log(`Calendar provider [${provider.provider}] registered.`);
   }
 
-  unregister(provider: CalendarProvider) {
+  unregister(provider: TProvider) {
     this.#providers.delete(provider.provider);
     this.logger.log(`Calendar provider [${provider.provider}] unregistered.`);
   }
