+++ packages/frontend/core/src/blocksuite/ai/components/page-editor-block-specs.ts
@@ -1,0 +1,22 @@
+import { ViewExtensionManager } from '@blocksuite/affine/ext-loader';
+import { getInternalViewExtensions } from '@blocksuite/affine/extensions/view';
+import { BlockViewIdentifier } from '@blocksuite/affine/std';
+import type { ExtensionType } from '@blocksuite/affine/store';
+import { literal } from 'lit/static-html.js';
+
+const manager = new ViewExtensionManager([...getInternalViewExtensions()]);
+const customPageEditorBlockSpecs: ExtensionType[] = [
+  ...manager.get('page'),
+  {
+    setup: di => {
+      di.override(
+        BlockViewIdentifier('affine:page'),
+        () => literal`affine-page-root`
+      );
+    },
+  },
+];
+
+export const getCustomPageEditorBlockSpecs = () => {
+  return customPageEditorBlockSpecs;
+};
