+++ packages/frontend/core/src/blocksuite/attachment-viewer/index.tsx
@@ -1,4 +1,7 @@
-import { ViewBody, ViewHeader } from '@affine/core/modules/workbench';
+import {
+  ViewBody,
+  ViewHeader,
+} from '@affine/core/modules/workbench/view/view-islands';
 
 import { AttachmentFallback, AttachmentPreviewErrorBoundary } from './error';
 import { PDFViewer } from './pdf/pdf-viewer';
