+++ packages/frontend/core/src/modules/doc/index.ts
@@ -9,7 +9,8 @@ export { DocsService } from './services/docs';
 import type { Framework } from '@toeverything/infra';
 
 import { WorkspaceDBService } from '../db/services/db';
-import { WorkspaceScope, WorkspaceService } from '../workspace';
+import { WorkspaceScope } from '../workspace/scopes/workspace';
+import { WorkspaceService } from '../workspace/services/workspace';
 import { Doc } from './entities/doc';
 import { DocRecord } from './entities/record';
 import { DocRecordList } from './entities/record-list';
